refactor ui testcases
